The View is Clean
This is for the Views and cleaning the UI from everything and checking Everything is working

+++ Application/Features/Courses/Models/CourseDTO.cs
@@ -9,7 +9,7 @@ namespace Application.Features.Courses.Models
 
         public int CreditHours;
 
-        public int CourseCategoryId;
+        public long CourseCategoryId;
 
 
     }
